myproject/requirements.txt
selenium
beautifulsoup4
requests
pillow
matplotlib